+++ lib/views/reports.dart
@@ -2,6 +2,7 @@ import 'package:alertabq/widgets/custom_drawer.dart';
 import 'package:alertabq/widgets/custom_navigation_bar.dart';
 import 'package:alertabq/widgets/pannic_button.dart';
 import 'package:alertabq/widgets/report_button.dart';
+import 'package:alertabq/widgets/report_card.dart';
 import 'package:flutter/material.dart';
 
 class Reports extends StatefulWidget {
@@ -55,11 +56,15 @@ class _ReportsState extends State<Reports> {
     Navigator.pushNamed(context, '/SubmitReport');
   }
 
+  void _onConfirm() {}
+  void _onNegative() {}
   @override
   Widget build(BuildContext context) {
     bool isDarkMode = Theme.of(context).brightness == Brightness.dark;
     Color textColor =
         isDarkMode ? const Color(0xFFF8F9FA) : const Color(0xFF2D3748);
+    String buttonAction = 'Confirmar';
+    String negativeButtonAction = 'No info';
     return Scaffold(
       appBar: AppBar(
         title: const Text('Reportes'),
@@ -97,14 +102,22 @@ class _ReportsState extends State<Reports> {
                           attachment: null,
                         ),
                       ],
+                      onConfirm: _onConfirm,
+                      buttonAction: buttonAction,
+                      onNegative: _onNegative,
+                      negativeButtonAction: negativeButtonAction,
                     ),
                     const SizedBox(height: 20),
-                    const ReportCard(
+                    ReportCard(
                       location: 'Ubicación 2',
                       dateTime: '2023-10-02 10:00',
                       description: 'Descripción del reporte 2',
                       attachment: null,
-                      events: [],
+                      events: const [],
+                      onConfirm: _onConfirm,
+                      buttonAction: buttonAction,
+                      onNegative: _onNegative,
+                      negativeButtonAction: negativeButtonAction,
                     ),
                   ],
                 ),
@@ -128,137 +141,4 @@ class _ReportsState extends State<Reports> {
       ),
     );
   }
-}
-
-class ReportCard extends StatefulWidget {
-  final String location;
-  final String dateTime;
-  final String description;
-  final String? attachment;
-  final List<Event> events;
-
-  const ReportCard({
-    required this.location,
-    required this.dateTime,
-    required this.description,
-    this.attachment,
-    required this.events,
-    super.key,
-  });
-
-  @override
-  State<ReportCard> createState() => _ReportCardState();
-}
-
-class _ReportCardState extends State<ReportCard> {
-  bool _isExpanded = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            if (widget.attachment != null)
-              Image.asset(
-                widget.attachment!,
-                width: 100,
-                height: 100,
-                fit: BoxFit.cover,
-              ),
-            const SizedBox(height: 10),
-            Text('Ubicación: ${widget.location}',
-                style: const TextStyle(fontWeight: FontWeight.bold)),
-            Text('Fecha y Hora: ${widget.dateTime}'),
-            Text('Descripción: ${widget.description}'),
-            const SizedBox(height: 10),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.end,
-              children: [
-                TextButton(
-                  onPressed: () {
-                    // Lógica para añadir eventos
-                  },
-                  child: const Text('Confirmar'),
-                ),
-              ],
-            ),
-            const Divider(),
-            GestureDetector(
-              onTap: () {
-                setState(() {
-                  _isExpanded = !_isExpanded;
-                });
-              },
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  const Text('Eventos',
-                      style: TextStyle(fontWeight: FontWeight.bold)),
-                  Icon(_isExpanded ? Icons.expand_less : Icons.expand_more),
-                ],
-              ),
-            ),
-            if (_isExpanded)
-              Column(
-                children: widget.events.map((event) {
-                  return Padding(
-                    padding: const EdgeInsets.symmetric(vertical: 8.0),
-                    child: EventCard(event: event),
-                  );
-                }).toList(),
-              ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class EventCard extends StatelessWidget {
-  final Event event;
-
-  const EventCard({required this.event, super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      color: Colors.grey[600],
-      child: Padding(
-        padding: const EdgeInsets.all(8.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            if (event.attachment != null)
-              Image.asset(
-                event.attachment!,
-                width: 50,
-                height: 50,
-                fit: BoxFit.cover,
-              ),
-            Text('Ubicación: ${event.location}',
-                style: const TextStyle(fontWeight: FontWeight.bold)),
-            Text('Fecha y Hora: ${event.dateTime}'),
-            Text('Descripción: ${event.description}'),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class Event {
-  final String location;
-  final String dateTime;
-  final String description;
-  final String? attachment;
-
-  Event({
-    required this.location,
-    required this.dateTime,
-    required this.description,
-    this.attachment,
-  });
 }
